SST: Use asset ID instead of handle in all APIs
This patch updates the SST interfaces to:
- use asset ID instead of asset handle as the object is always
  close at the end of each request
- change the asset ID type from uint16_t to uint32_t to allow
  longer asset IDs which can contain vendor specific ID codes
- remove tfm_sst_get_handle API
- updates all files which depend on these APIs and
  their documentation

+++ test/interface/include/sst_test_service_svc.h
@@ -33,7 +33,7 @@ enum tfm_sst_err_t sst_test_service_svc_setup(void);
  *
  * \return Returns error code as specified in \ref tfm_sst_err_t
  */
-enum tfm_sst_err_t sst_test_service_svc_dummy_encrypt(uint16_t key_uuid,
+enum tfm_sst_err_t sst_test_service_svc_dummy_encrypt(uint32_t key_uuid,
                                                       uint8_t *buf,
                                                       uint32_t buf_size);
 
@@ -47,7 +47,7 @@ enum tfm_sst_err_t sst_test_service_svc_dummy_encrypt(uint16_t key_uuid,
  *
  * \return Returns error code as specified in \ref tfm_sst_err_t
  */
-enum tfm_sst_err_t sst_test_service_svc_dummy_decrypt(uint16_t key_uuid,
+enum tfm_sst_err_t sst_test_service_svc_dummy_decrypt(uint32_t key_uuid,
                                                       uint8_t *buf,
                                                       uint32_t buf_size);
