+++ test/test_services/tfm_sst_test_service/sst_test_service_api.c
@@ -17,7 +17,7 @@ enum tfm_sst_err_t sst_test_service_setup(void)
 }
 
 __attribute__ ((naked))
-enum tfm_sst_err_t sst_test_service_dummy_encrypt(uint16_t key_uuid,
+enum tfm_sst_err_t sst_test_service_dummy_encrypt(uint32_t key_uuid,
                                                   uint8_t *buf,
                                                   uint32_t buf_size)
 {
@@ -26,7 +26,7 @@ enum tfm_sst_err_t sst_test_service_dummy_encrypt(uint16_t key_uuid,
 }
 
 __attribute__ ((naked))
-enum tfm_sst_err_t sst_test_service_dummy_decrypt(uint16_t key_uuid,
+enum tfm_sst_err_t sst_test_service_dummy_decrypt(uint32_t key_uuid,
                                                   uint8_t *buf,
                                                   uint32_t buf_size)
 {
