+++ test/test_services/tfm_sst_test_service/sst_test_service_svc.c
@@ -15,7 +15,7 @@ enum tfm_sst_err_t sst_test_service_svc_setup(void)
     return sst_test_service_veneer_setup();
 }
 
-enum tfm_sst_err_t sst_test_service_svc_dummy_encrypt(uint16_t key_uuid,
+enum tfm_sst_err_t sst_test_service_svc_dummy_encrypt(uint32_t key_uuid,
                                                       uint8_t *buf,
                                                       uint32_t buf_size)
 {
@@ -27,7 +27,7 @@ enum tfm_sst_err_t sst_test_service_svc_dummy_encrypt(uint16_t key_uuid,
                                                  buf_size);
 }
 
-enum tfm_sst_err_t sst_test_service_svc_dummy_decrypt(uint16_t key_uuid,
+enum tfm_sst_err_t sst_test_service_svc_dummy_decrypt(uint32_t key_uuid,
                                                       uint8_t *buf,
                                                       uint32_t buf_size)
 {
