+++ test/test_services/tfm_sst_test_service/sst_test_service_veneers.c
@@ -21,7 +21,7 @@ enum tfm_sst_err_t sst_test_service_veneer_setup(void)
 
 __tfm_secure_gateway_attributes__
 enum tfm_sst_err_t sst_test_service_veneer_dummy_encrypt(uint32_t app_id,
-                                                         uint16_t key_uuid,
+                                                         uint32_t key_uuid,
                                                          uint8_t *buf,
                                                          uint32_t buf_size)
 {
@@ -32,7 +32,7 @@ enum tfm_sst_err_t sst_test_service_veneer_dummy_encrypt(uint32_t app_id,
 
 __tfm_secure_gateway_attributes__
 enum tfm_sst_err_t sst_test_service_veneer_dummy_decrypt(uint32_t app_id,
-                                                         uint16_t key_uuid,
+                                                         uint32_t key_uuid,
                                                          uint8_t *buf,
                                                          uint32_t buf_size)
 {
